refacto: export default devices independently
