+++ example/src/App.tsx
@@ -9,7 +9,7 @@ ScreenSizer.setup();
 export default function App() {
   return (
     <SafeAreaProvider>
-      <ScreenSizer.Wrapper>
+      <ScreenSizer.Wrapper devices={ScreenSizer.defaultDevices.all}>
         <View style={styles.container}>
           <Text style={styles.title}>ScreenSizer Demo Application</Text>
           <Button
